+++ frontend/src/pages/teams/TeamList.jsx
@@ -1,4 +1,4 @@
-// beacon2/frontend/src/pages/groups/TeamList.jsx
+// beacon2/frontend/src/pages/teams/TeamList.jsx
 
 import { useState, useEffect, useRef } from 'react';
 import { Link, useNavigate } from 'react-router-dom';